+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 v0.19.0 (unreleased)
 ....................
 * Support ``Callable`` type hint, fix #279 by @proofit404
+* Fix schema for fields with ``validator`` decorator, fix #375 by @tiangolo
 
 v0.18.2 (2019-01-22)
 ....................